Update single_hero.php
this is what I have been working on. I am trying to get the join query to work right so that underneath each hero and their info, displays a picture of their creator and their name. I was also hoping to have a list underneath that of the series they are in, but I have not gotten that far yet. I just wanted ya'll to know im working but I havent gotten it figured out yet. This mysql query above does not work yet in the actual code.